Dev (#89) support for pydantic < 3.0
* enhanced intro in docs

* pydantic < 3.0

* python 3.12

* fix mnist example

* fix #87

* minor docs fixes

* add pydantic > 2 support

* add `setuptools` to reqs

* Revert "add `setuptools` to reqs"

This reverts commit a26e905b9d35f05a9050fd6eb2accb7c6e688567.

* ci tests for 3.12

* fix ci

* package build for 3.12

* if else fi

* false <-> true

* shitty ass ci matrix fix

* shitty ass ci matrix refix

* shitty ass ci matrix rerefix

* ci matrix fix

* ci matrix fix

* no torch 1.13 for python 3.11

* manually install lightning in ci

* sort imports

* some style

* a bit more style

* maybe somehow someway this should fix all my problems

* remove redundant lines

* empty

* upload-artifaccts@v4

* ci refire

* support for python >= 3.10 only

* python 3.13

* pytorch 2.6 prep

* python & torch ci versions fixed

* no exclude

* torch should be installed after the package?

* in accordance with release compatibility matrix

* py 3.13 is experimental for torch 2.5.0

* Revert "py 3.13 is experimental for torch 2.5.0"

This reverts commit 78f18771bd8a03aa2584208039a17089aad83d47.

* py 3.13 is experimental for torch 2.5.0

* torch.load with weights_only=False

* no more pytorch < 2

* pytorch 2.1.0 supports only numpy < 2

* fix coverage ci path

* update flake8 for lint

* update flake8 for lint

* fix numpy version adjusting

* little experiment

* no parenthesis

* `MetricMonitor` docs improved

* examples for CLI

---------

+++ thunder/cli/backend_cli.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import Dict, List, Union
 
 import yaml
 from rich.console import Console
@@ -7,6 +7,7 @@ from typer import Abort, Argument, Option, Typer
 from typing_extensions import Annotated
 
 from ..backend import MetaEntry
+from ..pydantic_compat import model_dump, model_validate
 from .backend import BACKENDS_CONFIG_PATH, BackendEntryConfig, load_backend_configs
 
 
@@ -20,6 +21,9 @@ BackendParamsArg = Annotated[List[str], Argument(
     show_default=False, help="Parameters of added run config.")]
 ForceAddArg = Annotated[bool, Option(
     "--force", "-f", help="Forces overwriting of the same backend in .yml file."
+)]
+CreateYMLArg = Annotated[bool, Option(
+    "--create", "-c", help="Creates .yml file with stored backends."
 )]
 
 console = Console()
@@ -35,13 +39,14 @@ def _set(name: BackendNameArg):
                       f"available configs: {sorted(local)}`")
         raise Abort(1)
 
-    local["meta"] = MetaEntry.parse_obj({"default": name})
+    local["meta"] = model_validate(MetaEntry, {"default": name})
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command()
-def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = False):
+def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = False,
+        create_yml: CreateYMLArg = False):
     """ Add run config to the list of available configs. """
     local = load_backend_configs()
     if name in local and not force:
@@ -52,9 +57,19 @@ def add(name: BackendNameArg, params: BackendParamsArg, force: ForceAddArg = Fal
     kwargs = dict(map(lambda p: p.split("="), params))
     config = {"backend": kwargs.pop("backend", "cli"), "config": kwargs}
 
-    local.update({name: BackendEntryConfig.parse_obj(config)})
+    local.update({name: model_validate(BackendEntryConfig, config)})
+
+    if not BACKENDS_CONFIG_PATH.parent.exists() and not create_yml:
+        path = str(BACKENDS_CONFIG_PATH)
+        console.print(f"Backends storage {path} does not exist, "
+                      f"you can create it by adding --create to the command.")
+        raise Abort(1)
+
+    if create_yml:
+        BACKENDS_CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
+
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command()
@@ -67,7 +82,7 @@ def remove(name: BackendNameArg):
 
     local.pop(name)
     with BACKENDS_CONFIG_PATH.open("w") as stream:
-        yaml.safe_dump({k: v.dict() for k, v in local.items()}, stream)
+        yaml.safe_dump({k: _dump_backend_entry(v) for k, v in local.items()}, stream)
 
 
 @backend_app.command(name="list")
@@ -78,15 +93,25 @@ def _list(names: BackendNamesArg = None):
     table = Table("Name", "Backend", "Parameters",
                   title=f"Configs at {str(BACKENDS_CONFIG_PATH.resolve())}")
 
+    if names is None:
+        names = local.copy()
+
     extra = set(names) - set(local)
     if extra:
         console.print("These names are not among your configs:", extra)
 
     for name in sorted(set(names if names else local) - extra.union({"meta"})):
-        entry = local[name].dict()
+        entry = _dump_backend_entry(local[name])
         table.add_row(*map(str, [name, entry.get("backend", None), entry.get("config", None)]))
 
     console.print(table)
 
     if "meta" in local:
         console.print(f"[italic green]Default is [/italic green]{local['meta'].default}")
+
+
+def _dump_backend_entry(backend: BackendEntryConfig) -> Dict[str, Union[str, Dict]]:
+    entry = model_dump(backend)
+    if hasattr(backend, "config"):
+        entry["config"] = model_dump(backend.config)
+    return entry
